requirements.txt
pytest~=5.2.1
python-daemon~=2.3.0